Moved user model to models folder
