+++ app/Models/Rentable.php
@@ -22,13 +22,13 @@ class Rentable extends Model
      */
     public function user()
     {
-        return $this->belongsTo('App\User', 'user_id');
+        return $this->belongsTo('App\Models\User', 'user_id');
     }
     /**
      * Get the lease for the place.
      */
     public function lease()
     {
-        return $this->hasMany('App\Lease');
+        return $this->hasMany('App\Models\Lease');
     }
 }
